Removed unnecessary files and fixed f-3-3 & f-1-2

+++ function-1-2.cpp
@@ -6,8 +6,8 @@ double array_mean(int array[], int n) {
     if (n < 1) {
         return 0.0;
     }
-    int sum = 0;
-    int mean = 0;
+    double sum = 0;
+    double mean = 0;
     for (int i = 0; i < n; i++) {
         sum = sum + array[i];
     }
