+++ function-3-3.cpp
@@ -7,7 +7,7 @@ double weighted_average(int array[], int n) {
     if (n < 1) {
         return 0;
     }
-    int occur = 0;
+    double occur = 0;
     double weightAvg = 0;
     for (int i = 0; i < n; i++) {
         occur = 0;
